refactor(auth): extract refresh logic
