+++ src/libs/auth/src/lib.rs
@@ -2,6 +2,7 @@ pub mod delegation;
 pub mod openid;
 pub mod profile;
 mod random;
+mod refresh;
 pub mod state;
 pub mod strategies;
