+++ src/libs/auth/src/openid/jwkset/mod.rs
@@ -1,4 +1,3 @@
-mod constants;
 mod fetch;
 mod impls;
 mod jwks;
