+++ src/libs/auth/src/refresh/types.rs
@@ -1,0 +1,6 @@
+pub enum RefreshStatus {
+    AllowedFirstFetch,
+    AllowedAfterCooldown,
+    AllowedRetry,
+    Denied,
+}
